sql: improve UNIQUE constraint checks
- Optimize UNIQUE constraint checks, remove additional uniqueness checks
  for new index entry in SQL if user didn't specified ON CONFLICT
  REPLACE/IGNORE/FAIL/ROLLBACK.
- Optimize UNIQUE checks if user specified one on error action, for
  example INSERT OR IGNORE, INSERT OR FAIL.
- Fix some test cases, because error messages had changed after moving
  checks from SQL to Tarantool.

Closes #2255

+++ src/box/sql/vdbe.h
@@ -63,7 +63,7 @@ struct VdbeOp {
 	u8 opcode;		/* What operation to perform */
 	signed char p4type;	/* One of the P4_xxx constants for p4 */
 	u8 notUsed1;
-	u8 p5;			/* Fifth parameter is an unsigned character */
+	u16 p5;			/* Fifth parameter is an unsigned character */
 	int p1;			/* First operand */
 	int p2;			/* Second parameter (often the jump destination) */
 	int p3;			/* The third parameter */
